+++ src/services/index.js
@@ -1,0 +1,4 @@
+const productServices = require('./productServices');
+module.exports={
+    productServices
+}
